# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SOCRA Parcours",
    "description": "Project for SOCRA",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/parcours/": {
      "get": {
        "tags": [
          "Parcours"
        ],
        "description": "Get all parcours ordered by creation date DESC",
        "parameters": [
          {
            "name": "campus",
            "in": "query",
            "description": "Filter search with campus name",
            "required": false,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter search with type of parcour",
            "required": false,
            "type": "string"
          },
          {
            "name": "cost",
            "in": "query",
            "description": "Filter search with filter with a lower price than the one given",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Parcours successfully obtained",
            "schema": {
              "$ref": "#/definitions/myParcours"
            }
          },
          "400": {
            "description": "Parcours successfully obtained"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Parcours"
        ],
        "description": "Get all parcours by keywords relevance",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "keywords": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parcours successfully obtained"
          },
          "400": {
            "description": "body/keywords are null"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    }
  },
  "definitions": {
    "myParcours": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "MTI"
          },
          "campus": {
            "type": "string",
            "example": "Paris"
          },
          "durantion": {
            "type": "number",
            "example": 24
          },
          "type": {
            "type": "string",
            "example": "Master"
          },
          "cost": {
            "type": "number",
            "example": 25000
          },
          "presential": {
            "type": "number",
            "example": 50
          },
          "start": {
            "type": "object",
            "properties": {}
          },
          "modules": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "S9"
                },
                "subjects": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "Projet IoT / Objet Connecté Médical OpenSource"
                      },
                      "description": {
                        "type": "string",
                        "example": "Initiation aux bases du projet IoT / Objet Connecté Médical OpenSource"
                      }
                    }
                  }
                }
              }
            }
          },
          "description": {
            "type": "string",
            "example": "La Majeure Santé a pour objectif de former de futurs ingénieurs en informatique pouvant s’interfacer efficacement avec des médecins et tous les professionnels du domaine de la santé.Les étudiants pourront ainsi donner du sens à leur carrière professionnelle en participant à la révolution des applications des données médicales et de l’IA pour la Santé comme, par exemple, l’analyse d’images médicales ou l’analyses de données génétiques ou le développement d’algorithmes pour mieux gérer les maladies des patients via des objets connectés.Du signal à la Data et de la Data à l’IA, les innovations numériques changent les habitudes des soignants et des patients et les ingénieurs font partie de ceux qui révolutionneront la santé de demain."
          }
        }
      }
    }
  }
}